đẩy code admin lên

+++ backend/src/main/java/com/cinema/modules/discount/repository/DiscountRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DiscountRepository extends JpaRepository<Discount, Integer> {
+public interface DiscountRepository extends JpaRepository<Discount, Long> {
     Optional<Discount> findByDiscountCode(String discountCode);
 }
